小程序20230531 第三版 购物车功能完成版

+++ src/main/java/com/Vno/controller/GoodsInfoController.java
@@ -95,7 +95,7 @@ public class GoodsInfoController {
 
     @ApiOperation("根据商品id查询商品")
     @GetMapping("/detail/{id}")
-    public Result<GoodsInfo> detail(@PathVariable Integer id){
+    public Result<GoodsInfo> detail(@PathVariable Long id){
         GoodsInfo byId = goodsInfoService.findById(id);
         return Result.success(byId);
     }
@@ -125,4 +125,6 @@ public class GoodsInfoController {
         return Result.success(page);
 
     }
+
+
 }
